refactor: auction status, item condition enum class를 constant로 이동 후 패키지 경로 수정
