--- src/main/java/com/samyookgoo/palgoosam/auction/domain/AuctionStatus.java
@@ -1,8 +1,0 @@
-package com.samyookgoo.palgoosam.auction.domain;
-
-public enum AuctionStatus {
-    pending,
-    active,
-    completed,
-    cancelled
-}
