--- src/main/java/com/samyookgoo/palgoosam/auction/domain/ItemCondition.java
@@ -1,9 +1,0 @@
-package com.samyookgoo.palgoosam.auction.domain;
-
-public enum ItemCondition {
-    brand_new,
-    like_new,
-    gently_used,
-    heavily_used,
-    damaged
-}
