+++ src/main/java/com/samyookgoo/palgoosam/auction/dto/response/AuctionUpdatePageResponse.java
@@ -1,7 +1,7 @@
 package com.samyookgoo.palgoosam.auction.dto.response;
 
-import com.samyookgoo.palgoosam.auction.domain.AuctionStatus;
-import com.samyookgoo.palgoosam.auction.domain.ItemCondition;
+import com.samyookgoo.palgoosam.auction.constant.AuctionStatus;
+import com.samyookgoo.palgoosam.auction.constant.ItemCondition;
 import java.util.List;
 import lombok.Builder;
 import lombok.Getter;
