+++ src/main/java/com/samyookgoo/palgoosam/auction/dto/response/AuctionUpdateResponse.java
@@ -1,6 +1,6 @@
 package com.samyookgoo.palgoosam.auction.dto.response;
 
-import com.samyookgoo.palgoosam.auction.domain.ItemCondition;
+import com.samyookgoo.palgoosam.auction.constant.ItemCondition;
 import java.util.List;
 import lombok.Builder;
 import lombok.Getter;
